add/fix: InfluxDB 파싱, 조회 등 실패시 로그 보강, 센서 종료시 cultivationSensor 소프트 삭제 및 룰엔진, DataGenerator에 삭제 명령 전송 구현, 테스트코드 추

+++ src/test/java/site/yesaido/cultivation_server/sensor/controller/SensorValueControllerTest.java
@@ -79,7 +79,7 @@ class SensorValueControllerTest {
         String sensorType = "TEMPERATURE";
         SensorTrendPointListResponse response = new SensorTrendPointListResponse(
                 CULTIVATION_ID, deviceEui, sensorType, "C",
-                List.of(new SensorTrendPointResponse(Instant.now(), 22.5))
+                List.of(new SensorTrendPointResponse(Instant.now(), BigDecimal.valueOf(22.5)))
         );
         given(influxService.findTrend(CULTIVATION_ID, deviceEui, sensorType)).willReturn(response);
